Extract population business logic from PopulationTask
