+++ src/world/generator/PopulationTask.php
@@ -27,8 +27,6 @@ use pocketmine\scheduler\AsyncTask;
 use pocketmine\utils\AssumptionFailedError;
 use pocketmine\world\format\Chunk;
 use pocketmine\world\format\io\FastChunkSerializer;
-use pocketmine\world\SimpleChunkManager;
-use pocketmine\world\World;
 use function array_map;
 use function igbinary_serialize;
 use function igbinary_unserialize;
@@ -71,8 +69,6 @@ class PopulationTask extends AsyncTask{
 		if($context === null){
 			throw new AssumptionFailedError("Generator context should have been initialized before any PopulationTask execution");
 		}
-		$generator = $context->getGenerator();
-		$manager = new SimpleChunkManager($context->getWorldMinY(), $context->getWorldMaxY());
 
 		$chunk = $this->chunk !== null ? FastChunkSerializer::deserializeTerrain($this->chunk) : null;
 
@@ -93,21 +89,15 @@ class PopulationTask extends AsyncTask{
 			$serialChunks
 		);
 
-		self::setOrGenerateChunk($manager, $generator, $this->chunkX, $this->chunkZ, $chunk);
-
-		$resultChunks = []; //this is just to keep phpstan's type inference happy
-		foreach($chunks as $relativeChunkHash => $c){
-			World::getXZ($relativeChunkHash, $relativeX, $relativeZ);
-			$resultChunks[$relativeChunkHash] = self::setOrGenerateChunk($manager, $generator, $this->chunkX + $relativeX, $this->chunkZ + $relativeZ, $c);
-		}
-		$chunks = $resultChunks;
-
-		$generator->populateChunk($manager, $this->chunkX, $this->chunkZ);
-		$chunk = $manager->getChunk($this->chunkX, $this->chunkZ);
-		if($chunk === null){
-			throw new AssumptionFailedError("We just generated this chunk, so it must exist");
-		}
-		$chunk->setPopulated();
+		[$chunk, $chunks] = PopulationUtils::populateChunkWithAdjacents(
+			$context->getWorldMinY(),
+			$context->getWorldMaxY(),
+			$context->getGenerator(),
+			$this->chunkX,
+			$this->chunkZ,
+			$chunk,
+			$chunks
+		);
 
 		$this->chunk = FastChunkSerializer::serializeTerrain($chunk);
 
@@ -116,18 +106,6 @@ class PopulationTask extends AsyncTask{
 			$serialChunks[$relativeChunkHash] = $c->isTerrainDirty() ? FastChunkSerializer::serializeTerrain($c) : null;
 		}
 		$this->adjacentChunks = igbinary_serialize($serialChunks) ?? throw new AssumptionFailedError("igbinary_serialize() returned null");
-	}
-
-	private static function setOrGenerateChunk(SimpleChunkManager $manager, Generator $generator, int $chunkX, int $chunkZ, ?Chunk $chunk) : Chunk{
-		$manager->setChunk($chunkX, $chunkZ, $chunk ?? new Chunk([], false));
-		if($chunk === null){
-			$generator->generateChunk($manager, $chunkX, $chunkZ);
-			$chunk = $manager->getChunk($chunkX, $chunkZ);
-			if($chunk === null){
-				throw new AssumptionFailedError("We just set this chunk, so it must exist");
-			}
-		}
-		return $chunk;
 	}
 
 	public function onCompletion() : void{
